Add custom board label color callback
